% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSort.R
\name{colSort}
\alias{colSort}
\title{Matrix Sort}
\usage{
colSort(X)
}
\arguments{
\item{X}{matrix to sort.}
}
\value{
Same matrix with rows permuted according to sort order.
}
\description{
Sorts a matrix by first column, breaking ties with second column, breaking those ties with 3rd, etc.
}
\keyword{internal}
