% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.format.R
\name{geno.format}
\alias{geno.format}
\title{Format genotype matrix}
\usage{
geno.format(X, Y.only = TRUE)
}
\arguments{
\item{X}{Genotype matrix or data.frame.}

\item{Y.only}{Logical.  Whether or not to output only \code{Y} (see return).}
}
\value{
A list with elements
\itemize{
\code{Y}: A 4-column numeric matrix.  Each row is sorted with zeros indicating missing alleles.
\code{A}: Vector of allele names.
\code{G}: Unique allele combinations, i.e. \code{Y} without duplicated rows.
}
}
\description{
Convert a 4-column genotype matrix or data.frame into a simplified numerical format.
}
\keyword{internal}
