% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.data.R
\name{HW.data}
\alias{HW.data}
\title{Convert Hardy-Weinberg sufficient statistics to \pkg{rstan} Input}
\usage{
HW.data(suff, debug = FALSE)
}
\arguments{
\item{suff}{Result of a call to \code{\link{HW.suff}}.}
}
\value{
a list with the following elements:
\itemize{
 \item \code{nG}: number of nonzero full genotype observations
 \item \code{nH}: number of nonzero inherited genotype observations
 \item \code{Xg}: counts in each unique genotype from observed data
 \item \code{Hg}: full genotype information - for each genotype, list all inherited genotypes \code{H} for each parent (unordered), at most 6 per genotype
 \item \code{nHg}: absolute value of \code{Hg} for each observed genotype
}
}
\description{
Convert Hardy-Weinberg sufficient statistics to \pkg{rstan} Input
}
\details{
The potential chromosomes in the observed data can produce more genotypes than are found in the observed data, in which case an observation with 0 counts is added to the group of all these unobserved counts.
}
\keyword{internal}

