% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.eqtest.R
\name{HW.eqtest}
\alias{HW.eqtest}
\title{Simulate test statistics for two samples from the same Hardy-Weinberg population model.}
\usage{
HW.eqtest(N1, N2, H, rho, nreps, verbose = TRUE, debug = FALSE)
}
\arguments{
\item{N1}{Size of sample 1.}

\item{N2}{Size of sample 2.}

\item{H}{2-column matrix of inheritable chromosomes.}

\item{rho}{Vector of probabilities for inheritance of each chromosome from either parent.  Can also be a matrix, in which case each simulation randomly draws with replacement from the rows of rho.}

\item{nreps}{Number of replications of the simulation.}
}
\value{
An \code{nreps x 2} matrix with the simulated chi-squared values in the first column and the simulated LRT values in the second.
}
\description{
Samples chi-squared and likelihood ratio test statistics two samples from an multinomial distribution constrained to satisfy the Hardy-Weinberg equilibrium.
}
\keyword{internal}

