% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.post.R
\name{HW.post}
\alias{HW.post}
\title{Posterior inference for Hardy-Weinberg model}
\usage{
HW.post(nsamples, X, full.stan.out = FALSE, ...)
}
\arguments{
\item{nsamples}{Number of posterior samples}

\item{X}{4-column matrix of observations in the correct format.}

\item{full.stan.out}{Logical.  Whether or not to return the full stan output.  For monitoring convergence of the MCMC sampling.}

\item{...}{Further arguments to be passed to Stan.}
}
\value{
A list with elements
\itemize{
  \item \code{A}: The unique allele names.
  \item \code{H}: The 2-column matrix potential chromosomes.
  \item \code{rho}: A matrix with \code{ncol(rho) == nrow(H)}, where each row is a draw from the posterior distribution of inheritance probabilities.
  \item \code{sfit}: If \code{full.stan.out = TRUE}, the fitted \code{stan} object.
}
}
\description{
MCMC sampling from the posterior parameter distribution of a Hardy-Weinberg model using \pkg{rstan}.
}
\keyword{internal}

