% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.suff.R
\name{HW.suff}
\alias{HW.suff}
\title{Sufficient statistics for the Hardy-Weinberg model}
\usage{
HW.suff(X, popId, debug = FALSE)
}
\arguments{
\item{X}{Genotype adata.  Either a \code{N x 4} matrix with \code{NA}'s indicating duplicates or a \code{N x 5} column data.frame with the first column being the \code{popId}.}

\item{popId}{grouping variable for \code{X}. Must be supplied if \code{X} has 4 columns.}
}
\value{
A list with elements:
\itemize{
\item \code{A}: Vector of unique alleles names.  The allele numbers in the following quantities correspond to the indices of \code{A}.
\item \code{G}: 4-column matrix of unique genotype combinations.  The presence of 0's indicates that there were less that a given genotype either has less than 4 distinct alleles or that some alleles are duplicated.
\item \code{Xg}: Vector of counts in each unique genotype combination from \code{X}.
\item \code{H}: 2-column matrix of all possible chromosomes that could have generated \code{X}.  Each chromosome consists of 2 alleles.
\item \code{Y}: Observed data in a simplified numerical format.  That is, all allele labels are given unique positive-integer identifiers and missing or duplicated alleles are represented by 0's.
}
}
\description{
Converts a matrix of genotype data into the sufficient statistics required to fit a  Hardy-Weinberg model.
}
\examples{
# Two fish: one with 2 alleles and one with 3
X <- rbind(c("a", "b", NA, NA), c("c", "a", "d", NA))
suff <- MADPop:::HW.suff(X)
suff$A # representing all of the unique alleles and 4 is repeated
suff$G # since there are only 2 unique full genotypes
suff$Xg # since there is only 1 of each genotype
suff$H # all possible chromosomes compatible with the observed data
suff$Y # dataset in simplified numerical format
}
\keyword{internal}

