% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.stat.R
\name{chi2.stat}
\alias{chi2.stat}
\title{Chi-squared test statistic for contingency tables}
\usage{
chi2.stat(tab)
}
\arguments{
\item{tab}{A \code{K x N} matrix (contingency table) of counts. See details.}
}
\value{
The calculated value of the chi-squared statistic.
}
\description{
Calculates the chi-squared test statistic for general two-way contingency tables.
}
\details{
The chi-squared test statistic is computed as
\deqn{\sum_{i=1}^K \sum_{j=1}^N (E_{ij} - O_{ij})^2/E_{ij},}
where \eqn{O_{ij} = \code{tab[i,j]}} and
\deqn{E_{ij} = \code{sum(tab[,i])*sum(tab[,j])/sum(tab)}.}
That is, the null hypothesis is that each row of the contingency table is drawn from the same multinomial distribution.  If any column has only zeros it is removed before calculating the test statistic.
}
\examples{
# simple contingency table
ctable <- rbind(pop1 = c(5, 3, 0, 3),
                pop2 = c(4, 10, 2, 5))
colnames(ctable) <- LETTERS[1:4]
ctable
chi2.stat(ctable) # chi^2 statistic
}

