% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hUM.post.R
\name{hUM.post}
\alias{hUM.post}
\title{Posterior draws from a fitted hierarchical Unconstrained-Multinomial model}
\usage{
hUM.post(nsamples, X, popId, rhoId, full.stan.out = FALSE, ...,
  debug = FALSE)
}
\arguments{
\item{nsamples}{Number of posterior samples}

\item{X}{4-column or 5-column matrix of observations in the correct format.  See \code{\link{UM.suff}}.}

\item{popId}{Optional vector of population identifiers.  See \code{\link{UM.suff}}.}

\item{rhoId}{Populations for which rho probabilities are desired.  Defaults to all populations.  Set \code{rhoId = NULL} not to compute these for any populations.}

\item{full.stan.out}{Logical.  Whether or not to return the full \code{stan} output.  For monitoring convergence of the MCMC sampling.}

\item{...}{Further arguments to be passed to \code{stan}.}
}
\value{
A list with elements
\itemize{
  \item \code{A}: The unique allele names.
  \item \code{G}: The 4-column matrix of unique genotype combinations.
  \item \code{rho}: A matrix with \code{ncol(rho) == nrow(G)}, where each row is a draw from the posterior distribution of inheritance probabilities.
  \item \code{sfit}: If \code{full.stan.out = TRUE}, the fitted \code{stan} object.
}
}
\description{
MCMC sampling from a Dirichlet-Multinomial model using \code{\link[rstan]{stan}}.
}
\details{
The model is given by
\deqn{Y_k \sim_{ind} \emph{Multin}(\rho_k, N_k), \qquad \rho_k \sim_{iid} \emph{Dir}(\alpha), \qquad (\bar \alpha, \alpha_0) \sim 1/(1+\alpha_0)^2.}
}
\examples{
# fit hierarchical model to fish215 data

# only output posterior samples for lake Simcoe
rhoId <- "Simcoe"
nsamples <- 500
hUM.fit <- hUM.post(nsamples = nsamples, X = fish215,
                    rhoId = rhoId,
                    chains = 1) # number of MCMC chains
rho.post <- hUM.fit$rho[,1,] # convert 3-d array to 2-d matrix

# plot first 20 posterior probabilities in lake Simcoe
boxplot(rho.post[,1:20], las = 2,
        xlab = "Genotype", ylab = "Posterior Probability",
        pch = ".", col = "grey")
}

