% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW.eqtest.R
\name{HW.eqtest}
\alias{HW.eqtest}
\title{Equality tests for two Hardy-Weinberg multinomial samples}
\usage{
HW.eqtest(N1, N2, H, rho, nreps, verbose = TRUE, debug = FALSE)
}
\arguments{
\item{N1, N2}{Sample sizes.}

\item{H}{2-column matrix of inheritable chromosomes.}

\item{rho}{Vector of probabilities for inheritance of each chromosome from either parent.  Can also be a matrix, in which case each simulation randomly draws with replacement from the rows of rho.}

\item{nreps}{Number of replications of the simulation.}
}
\value{
An \code{nreps x 2} matrix with the simulated chi-squared and LR values.
}
\description{
Generate multinomial samples from a common Hardy-Weinberg equilibrium distribution and calculate the Chi-square and Likelihood Ratio test statistics.
}
\keyword{internal}

