% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRT.stat.R
\name{LRT.stat}
\alias{LRT.stat}
\title{Likelihood ratio test statistic for contingency tables}
\usage{
LRT.stat(tab)
}
\arguments{
\item{tab}{A \code{K x N} matrix (contingency table) of counts. See details.}
}
\value{
The calculated value of the LRT statistic.
}
\description{
Calculate the likelihood ratio test statistic for general two-way contingency tables.
}
\details{
The likelihood ratio test statistic is computed as
\deqn{2 \sum_{i=1}^K \sum_{j=1}^N O_{ij} \log(p^A_{ij}/p^0_{j}),}
where \eqn{O_{ij} = \code{tab[i,j]}}, \eqn{p^A_{ij} = \code{tab[i,j]/sum(tab)}}, and
\deqn{p^0_j = \code{sum(tab[,j])/sum(tab)}.}
That is, the null hypothesis is that each row of the contingency table is drawn from the same multinomial distribution.  If any column has only zeros it is removed before calculating the test statistic.
}
\examples{
# simple contingency table
ctable <- rbind(pop1 = c(5, 3, 0, 3),
                pop2 = c(4, 10, 2, 5))
colnames(ctable) <- LETTERS[1:4]
ctable
LRT.stat(ctable) # likelihood ratio statistic
}

